+++ app/src/main/java/gr/eduinvoice/NavigationRoutes.kt
@@ -5,6 +5,7 @@ sealed class Screen(val route: String) {
     object Home : Screen("home")
     object Login : Screen("login")
     object Register : Screen("register")
+    object ResetPassword : Screen("resetPassword")
     object Students : Screen("students")
     object Student : Screen("student/{studentId}") {
         fun createRoute(studentId: Long) = "student/$studentId"